+++ photonact/__init__.py
@@ -1,13 +1,14 @@
 """PhotonAct public API."""
 
-from photonact.activations import CurveActivation
+from photonact.activations import CurveActivation, HysteresisActivation
 from photonact.curves import CurveData, CurveMetadata, load_curve
 
 __all__ = [
     "CurveActivation",
+    "HysteresisActivation",
     "CurveData",
     "CurveMetadata",
     "load_curve",
 ]
 
-__version__ = "0.0.1"
+__version__ = "0.1.0.dev0"